eliminate_duplicate_disjuncts(): Return deleted disjunct count
No need to return the disjunct list because the deletion is implemented
"in place", and the first disjunct is never deleted. Thus the given
argument of disjunct list still points to the list after duplicate
removal. Returning the count is more useful since it, e.g., enables to
find out the total disjuncts removed in the whole sentence.