Add macOS CI
Also do CI testing on macOS 12, 13, and 14. Not using "macos-latest"
because that currently refers to macOS 12 which is not the latest
version. It is good to test on multiple versions and on multiple
architectures. The macOS 12 and 13 runners are x86_64 while the macOS 14
runner is arm64. This list can be updated annually as new macOS versions
are released and old versions are removed from CI, or switched to
"macos-latest" once that actually points to the latest version again.

Do not fail fast; this ensures that a build failure on one runner does
not cancel the other runners so that temporarily expected failures on
one runner do not mask unexpected failures on other runners.

Use Homebrew (pre-installed on macOS runners) to install dependencies
for macOS.

Explicitly mention dependencies that are used, like autoconf, automake,
and libtool, even if they are in the set of dependencies already
pre-installed on the runners, because who knows if that will always be
the case.

Fixes #1496