Remove Java 64-bit macOS statement from README
As of Mac OS X 10.6 (released 2009), compilers build 64-bit by default
on 64-bit systems, and 10.6 is the last OS version that runs on 32-bit
systems. Naturally, users need to install a Java that matches the
architecture they want to build for, and as with any other software they
should supply -arch flags in CFLAGS, CXXFLAGS, and LDFLAGS if they
want to build for a non-default architecture.