Remove attempt to force x86_64 Java build on macOS
This code never took effect because the apple_osx variable on which is
was based was always "no". The claim that modern 64-bit Macs build
32-bit by default is false; compilers on macOS have built 64-bit by
default on 64-bit systems ever since Mac OS X 10.6 was released (2009).
There are also other 64-bit architectures than x86_64: current Macs use
arm64 while Macs of decades past used ppc64.

Mac OS X 10.6 was the last time Java was included with macOS; since
then, users have been expected to install a Java distribution of their
choosing -- and there are many to choose from, ever since Oracle made
the license for their Java distribution less appealing. It's the user's
responsibility to select a Java distribution that matches the
architecture they want to build for, and to specify -arch flags in
CFLAGS, CXXFLAGS, and LDFLAGS if that architecture is not their
compiler's default. No assumptions about this should be made by build
systems.