form_match_list(): Don't use _Atomic
There is no need to ensure the "id" variable is incremented in an atomic
way, because we don't care if two threads will use the same value.

The crash that got debugged on commit 9f73b372 was due to another
reason: not using "id" variable if it is 0. This has been fixed.

Removing it gets rid of some small overhead, and also allows MSVC compilation.